refactor: remove shared-users which was only used by back-office #74580
