+++ apps/back-office/src/app/dashboard/pages/users/graphql/queries.ts
@@ -1,6 +1,5 @@
 import { gql } from 'apollo-angular';
 
-// === GET ROLES ===
 /** Graphql query for getting roles (of an application or all) */
 export const GET_ROLES = gql`
   query GetRoles($all: Boolean, $application: ID) {
@@ -19,7 +18,6 @@ export const GET_ROLES = gql`
   }
 `;
 
-// === GET USERS ===
 /** Graphql query for getting users */
 export const GET_USERS = gql`
   {
