+++ apps/back-office/src/app/dashboard/pages/users/users.module.ts
@@ -2,12 +2,34 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { UsersRoutingModule } from './users-routing.module';
 import { UsersComponent } from './users.component';
-import { UsersModule as SharedUsersModule } from '@oort-front/shared';
+import { FilterComponent } from './filter/filter.component';
+import {
+  ButtonModule,
+  CheckboxModule,
+  IconModule,
+  MenuModule,
+  TableModule,
+  TooltipModule,
+} from '@oort-front/ui';
+import { TranslateModule } from '@ngx-translate/core';
+import { SkeletonTableModule } from '@oort-front/shared';
 
 /** Users page module */
 @NgModule({
   declarations: [UsersComponent],
-  imports: [CommonModule, UsersRoutingModule, SharedUsersModule],
+  imports: [
+    CommonModule,
+    UsersRoutingModule,
+    FilterComponent,
+    ButtonModule,
+    TranslateModule,
+    MenuModule,
+    TableModule,
+    CheckboxModule,
+    IconModule,
+    SkeletonTableModule,
+    TooltipModule,
+  ],
   exports: [UsersComponent],
 })
 export class UsersModule {}
