+++ libs/shared/src/lib/components/users/add-user/graphql/queries.ts
@@ -1,0 +1,20 @@
+import { gql } from 'apollo-angular';
+
+/** Graphql query for getting users */
+export const GET_USERS = gql`
+  {
+    users {
+      id
+      username
+      name
+      roles {
+        id
+        title
+        application {
+          id
+        }
+      }
+      oid
+    }
+  }
+`;
