--- libs/shared/src/lib/components/users/components/invite-users/invite-users.component.spec.ts
@@ -1,59 +1,0 @@
-import { HttpClientModule } from '@angular/common/http';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { UntypedFormBuilder } from '@angular/forms';
-import {
-  DialogModule as DialogCdkModule,
-  DialogRef,
-  DIALOG_DATA,
-} from '@angular/cdk/dialog';
-import {
-  TranslateModule,
-  TranslateService,
-  TranslateFakeLoader,
-  TranslateLoader,
-} from '@ngx-translate/core';
-
-import { InviteUsersComponent } from './invite-users.component';
-
-describe('InviteUsersComponent', () => {
-  let component: InviteUsersComponent;
-  let fixture: ComponentFixture<InviteUsersComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      providers: [
-        { provide: 'environment', useValue: {} },
-        UntypedFormBuilder,
-        { provide: DialogRef, useValue: {} },
-        {
-          provide: DIALOG_DATA,
-          useValue: {
-            access: { canSee: null, canUpdate: null, canDelete: null },
-          },
-        },
-        TranslateService,
-      ],
-      declarations: [InviteUsersComponent],
-      imports: [
-        HttpClientModule,
-        DialogCdkModule,
-        TranslateModule.forRoot({
-          loader: {
-            provide: TranslateLoader,
-            useClass: TranslateFakeLoader,
-          },
-        }),
-      ],
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(InviteUsersComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
