+++ libs/shared/src/lib/components/users/invite-users-modal/invite-users-modal.component.ts
@@ -1,15 +1,25 @@
 import { Component, Inject, ViewChild } from '@angular/core';
 import { DialogRef, DIALOG_DATA, Dialog } from '@angular/cdk/dialog';
 import { GridComponent, GridDataResult } from '@progress/kendo-angular-grid';
-import { Role, User } from '../../../../models/user.model';
-import { PositionAttributeCategory } from '../../../../models/position-attribute-category.model';
+import { Role, User } from '../../../models/user.model';
+import { PositionAttributeCategory } from '../../../models/position-attribute-category.model';
 import { FormBuilder, UntypedFormArray, Validators } from '@angular/forms';
-import { DownloadService } from '../../../../services/download/download.service';
+import { DownloadService } from '../../../services/download/download.service';
 import { TranslateService } from '@ngx-translate/core';
 import { UploadEvent } from '@progress/kendo-angular-upload';
 import { SnackbarService } from '@oort-front/ui';
 import { takeUntil } from 'rxjs';
-import { UnsubscribeComponent } from '../../../utils/unsubscribe/unsubscribe.component';
+import { UnsubscribeComponent } from '../../utils/unsubscribe/unsubscribe.component';
+import { CommonModule } from '@angular/common';
+import { GridModule, GroupModule } from '@progress/kendo-angular-grid';
+import { DropDownsModule } from '@progress/kendo-angular-dropdowns';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { ButtonModule, ButtonsModule } from '@progress/kendo-angular-buttons';
+import { AddUserModule } from '../add-user/add-user.module';
+import { TranslateModule } from '@ngx-translate/core';
+import { UploadsModule } from '@progress/kendo-angular-upload';
+import { ButtonModule as uiButtonModule, TextareaModule } from '@oort-front/ui';
+import { DialogModule } from '@oort-front/ui';
 
 /** Model fot the input data */
 interface DialogData {
@@ -20,13 +30,30 @@ interface DialogData {
   downloadPath: string;
 }
 
-/** Component for inviting users */
+/** Invite users modal component */
 @Component({
-  selector: 'shared-invite-users',
-  templateUrl: './invite-users.component.html',
-  styleUrls: ['./invite-users.component.scss'],
+  standalone: true,
+  selector: 'shared-invite-users-modal',
+  templateUrl: './invite-users-modal.component.html',
+  styleUrls: ['./invite-users-modal.component.scss'],
+  imports: [
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    GridModule,
+    GroupModule,
+    DropDownsModule,
+    ButtonModule,
+    ButtonsModule,
+    AddUserModule,
+    TranslateModule,
+    UploadsModule,
+    DialogModule,
+    TextareaModule,
+    uiButtonModule,
+  ],
 })
-export class InviteUsersComponent extends UnsubscribeComponent {
+export class InviteUsersModalComponent extends UnsubscribeComponent {
   public gridData: GridDataResult = { data: [], total: 0 };
   public formGroup!: ReturnType<typeof this.createFormGroup>;
   /** File size limit, in MB */
@@ -46,15 +73,15 @@ export class InviteUsersComponent extends UnsubscribeComponent {
   }
 
   /**
-   * Constructor of the component
+   * Invite users modal component
    *
-   * @param downloadService The download service
-   * @param snackBar The snack bar service
-   * @param fb The form builder service
-   * @param dialog The Dialog service
-   * @param dialogRef The reference to a Dialog
-   * @param translate The translation service
-   * @param data The input data of the component
+   * @param downloadService Shared download service
+   * @param snackBar Shared snack bar service
+   * @param fb Angular form builder service
+   * @param dialog CDK Dialog service
+   * @param dialogRef Dialog reference
+   * @param translate Angular translation service
+   * @param data Passed data
    * @param environment environment
    */
   constructor(
@@ -62,7 +89,7 @@ export class InviteUsersComponent extends UnsubscribeComponent {
     private snackBar: SnackbarService,
     private fb: FormBuilder,
     public dialog: Dialog,
-    public dialogRef: DialogRef<InviteUsersComponent>,
+    public dialogRef: DialogRef<InviteUsersModalComponent>,
     public translate: TranslateService,
     @Inject(DIALOG_DATA) public data: DialogData,
     @Inject('environment') public environment: any
