+++ libs/shared/src/lib/components/users/public-api.ts
@@ -1,2 +1,1 @@
-export * from './users.component';
-export * from './users.module';
+export * from './invite-users-modal/invite-users-modal.component';
