--- libs/shared/src/lib/components/users/users.component.html
@@ -1,172 +1,0 @@
-<div class="mb-8 flex flex-col gap-2">
-  <div class="flex gap-8 justify-between flex-col sm:flex-row">
-    <app-users-filter
-      #filter
-      (filter)="applyFilter($event)"
-      [loading]="loading"
-      [roles]="roles"
-    >
-    </app-users-filter>
-    <!-- Actions -->
-    <div class="flex gap-2">
-      <!-- Actions when selected items -->
-      <ng-container *ngIf="selection.selected.length > 0">
-        <ui-button
-          icon="delete"
-          category="secondary"
-          variant="danger"
-          (click)="onDelete(selection.selected)"
-        >
-          {{
-            'common.deleteObject'
-              | translate : { name: 'common.user.few' | translate }
-          }}
-        </ui-button>
-      </ng-container>
-
-      <!-- Default actions -->
-      <ui-button
-        icon="file_download"
-        category="secondary"
-        variant="primary"
-        [uiMenuTriggerFor]="menu"
-      >
-        {{
-          'common.exportObject'
-            | translate : { name: 'common.user.few' | translate }
-        }}
-      </ui-button>
-      <ui-menu #menu>
-        <button uiMenuItem (click)="onExport('csv')">.csv</button>
-        <button uiMenuItem (click)="onExport('xlsx')">.xlsx</button>
-      </ui-menu>
-      <ui-button
-        icon="send"
-        category="secondary"
-        variant="primary"
-        (click)="onInvite()"
-      >
-        {{ 'components.users.invite.confirm' | translate }}
-      </ui-button>
-    </div>
-  </div>
-  <ng-container *ngTemplateOutlet="filter.expandedFilter"></ng-container>
-</div>
-<!-- Table container -->
-<div class="overflow-x-hidden shadow-2lg">
-  <!-- Table scroll container -->
-  <div class="overflow-x-auto">
-    <table
-      *ngIf="!loading"
-      cdk-table
-      uiTableWrapper
-      [dataSource]="filteredUsers"
-    >
-      <!-- Checkbox Column -->
-      <ng-container cdkColumnDef="select">
-        <th uiCellHeader *cdkHeaderCellDef scope="col">
-          <ui-checkbox
-            (change)="$event ? masterToggle() : null"
-            [checked]="selection.hasValue() && isAllSelected()"
-            [indeterminate]="selection.hasValue() && !isAllSelected()"
-            [ariaLabel]="checkboxLabel()"
-          >
-          </ui-checkbox>
-        </th>
-        <td uiCell *cdkCellDef="let row">
-          <ui-checkbox
-            (click)="$event.stopPropagation()"
-            (change)="$event ? selection.toggle(row) : null"
-            [checked]="selection.isSelected(row)"
-            [ariaLabel]="checkboxLabel(row)"
-          >
-          </ui-checkbox>
-        </td>
-      </ng-container>
-      <ng-container cdkColumnDef="username">
-        <th uiCellHeader *cdkHeaderCellDef scope="col">
-          <span class="headerTitle">
-            {{ 'models.user.username' | translate }}
-          </span>
-        </th>
-        <td uiCell *cdkCellDef="let element">{{ element.username }}</td>
-      </ng-container>
-
-      <ng-container cdkColumnDef="name">
-        <th uiCellHeader *cdkHeaderCellDef scope="col">
-          <span class="headerTitle">
-            {{ 'common.name' | translate }}
-          </span>
-        </th>
-        <td
-          uiCell
-          *cdkCellDef="let element"
-          class="!text-gray-900 !font-medium"
-        >
-          {{ element.name }}
-        </td>
-      </ng-container>
-
-      <ng-container cdkColumnDef="roles">
-        <th uiCellHeader *cdkHeaderCellDef scope="col">
-          <span class="headerTitle">
-            {{ 'common.role.few' | translate }}
-          </span>
-        </th>
-        <td uiCell *cdkCellDef="let element" class="max-w-[30vw]">
-          <span *ngFor="let role of element.roles"> {{ role.title }}</span>
-        </td>
-      </ng-container>
-
-      <ng-container cdkColumnDef="oid">
-        <th uiCellHeader *cdkHeaderCellDef scope="col">
-          <span class="headerTitle">
-            {{ 'common.id' | translate }}
-          </span>
-        </th>
-        <td uiCell *cdkCellDef="let element">{{ element.oid }}</td>
-      </ng-container>
-
-      <ng-container cdkColumnDef="actions" [stickyEnd]="true">
-        <th uiCellHeader *cdkHeaderCellDef scope="col" class="w-16"></th>
-        <td uiCell *cdkCellDef="let element">
-          <ui-button
-            [isIcon]="true"
-            icon="more_vert"
-            [uiMenuTriggerFor]="menu"
-            (click)="$event.stopPropagation()"
-            [uiTooltip]="'common.actions' | translate"
-          >
-          </ui-button>
-          <ui-menu #menu>
-            <button uiMenuItem (click)="onDelete([element])">
-              <ui-icon icon="delete" variant="danger"></ui-icon>
-              {{ 'common.delete' | translate }}
-            </button>
-          </ui-menu>
-        </td>
-      </ng-container>
-
-      <tr cdk-header-row *cdkHeaderRowDef="displayedColumns"></tr>
-      <tr
-        cdk-row
-        *cdkRowDef="let row; columns: displayedColumns"
-        (click)="onClick(row)"
-        class="cursor-pointer"
-      ></tr>
-    </table>
-    <!-- Loading indicator -->
-    <shared-skeleton-table
-      *ngIf="loading"
-      [columns]="[
-        'common.name',
-        'models.user.username',
-        'common.id',
-        'common.role.one'
-      ]"
-      [actions]="true"
-      [checkbox]="true"
-    >
-    </shared-skeleton-table>
-  </div>
-</div>
