+++ libs/shared/src/lib/views/application-users/application-users.module.ts
@@ -5,7 +5,6 @@ import { ApplicationUsersRoutingModule } from './application-users-routing.modul
 import { UserListModule } from './components/user-list/user-list.module';
 import { TranslateModule } from '@ngx-translate/core';
 import { TabsModule, ButtonModule, MenuModule } from '@oort-front/ui';
-import { InviteUsersModule } from '../../components/users/components/invite-users/invite-users.module';
 
 /**
  * Application users view module.
@@ -19,7 +18,6 @@ import { InviteUsersModule } from '../../components/users/components/invite-user
     TranslateModule,
     TabsModule,
     MenuModule,
-    InviteUsersModule,
     ButtonModule,
   ],
   exports: [ApplicationUsersComponent],
